requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
lxml>=4.9.0
requests-html>=0.10.0
